histo.pl: inline histogram-ish producer
Useful for when you're slamming a bunch of stuff through egreps and similar
and want to summarize it at the end.  You simply give it the bucket size and
it eats the numbers and outputs a histogram-like thing (it doesn't output
empty buckets, so I think it mayn't be a "real" histogram).  Works for
integers and floats.  Also accepts dates, although it's pretty particular
about their format.  It auto-magically tries to recognize what kind of data
it's getting by running regexes (it is Perl after all) against the data.

Something like this

$ perl -e 'for (my $i = 0; $i < 100; $i++) { my $scale = 100; { my $x = int($scale * rand()); print "$x\n"; }}' | histo.pl
Bucket size --> 10
[ 0] for pattern '^\s*\d+(?:.\d+)?$' with line --> '61'
Matched histo of numbers
Size of array 100
Mean of 100 values is 53.37
Standard deviation is 25.38
Min is 1
Max is 97
0 6
10 5
20 9
30 10
40 12
50 14
60 15
70 11
80 8
90 10
--------------------------------------------------------------------------------